requirements.txt
streamlit
torch
torchvision
Pillow
gdown
numpy

